fix: extract functions to fix lint violations in top/index.ts
